requirements.txt
rembg
flask
flask-cors
gunicorn
pillow
